+++ packages/rust-bindings/src/external_functions.rs
@@ -345,3 +345,11 @@ pub fn generate_baker_keys_ext(sender: Base58String) -> JsonString {
     };
     error_to_string(generate_baker_keys(sender))
 }
+
+#[wasm_bindgen(js_name = deserializeTypeValue)]
+pub fn deserialize_type_value_ext(
+    serialized_value: JsonString,
+    value_type: HexString
+) -> Result<String, String> {
+    deserialize_type_value_aux(serialized_value, value_type).map_err(|e| format!("Unable to deserialize value due to: {}", e))
+}